refactor: Extracted embedded tests into separate classes
